reducer 제거 및 테스트 slice에 적용 및 리펙토링
